[feat]: Refactored packages, added files for each method. Removed ginkgo, added testify
