+++ internal/handling/error_test.go
@@ -4,72 +4,60 @@ import (
 	"testing"
 
 	"github.com/brianvoe/gofakeit/v7"
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"github.com/upassed/upassed-account-service/internal/handling"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
-var _ = Describe("Handling Application Errors Tests", func() {
-	Describe("Convert Application Error", func() {
-		It("should convert ApplicationError and add a custom additional details", func() {
-			message := gofakeit.Error().Error()
-			code := codes.Internal
-			applicationError := handling.NewApplicationError(message, code)
+func TestConvertApplicationError_AddDetails(t *testing.T) {
+	message := gofakeit.Error().Error()
+	code := codes.Internal
+	applicationError := handling.NewApplicationError(message, code)
 
-			handledError := handling.HandleApplicationError(applicationError)
+	handledError := handling.HandleApplicationError(applicationError)
 
-			st := status.Convert(handledError)
-			Expect(st.Code()).To(Equal(code))
-			Expect(st.Message()).To(Equal(message))
-			Expect(st.Details()).To(HaveLen(1))
-		})
+	st := status.Convert(handledError)
+	assert.Equal(t, code, st.Code())
+	assert.Equal(t, message, st.Message())
+	assert.Equal(t, 1, len(st.Details()))
+}
 
-		It("should wrap an application error with wrap options", func() {
-			message := "error message"
-			code := codes.AlreadyExists
+func TestConvertApplicationError_WrapOptions(t *testing.T) {
+	message := "error message"
+	code := codes.AlreadyExists
 
-			applicationError := handling.NewApplicationError(message, code)
-			wrappedError := handling.HandleApplicationError(applicationError, handling.WithCode(codes.OK))
+	applicationError := handling.NewApplicationError(message, code)
+	wrappedError := handling.HandleApplicationError(applicationError, handling.WithCode(codes.OK))
 
-			st := status.Convert(wrappedError)
+	st := status.Convert(wrappedError)
+	require.NotNil(t, wrappedError)
 
-			Expect(wrappedError).NotTo(BeNil())
-			Expect(st.Message()).To(Equal(message))
-			Expect(st.Code()).To(Equal(code))
-		})
-	})
+	assert.Equal(t, message, st.Message())
+	assert.Equal(t, code, st.Code())
+}
 
-	Describe("Convert not an Application Error", func() {
-		It("should wrap not an ApplicationError and add a custom additional details", func() {
-			initialError := gofakeit.Error()
+func TestConvertApplicationError_WrappingNotAnApplicationError(t *testing.T) {
+	initialError := gofakeit.Error()
 
-			handledError := handling.HandleApplicationError(initialError)
+	handledError := handling.HandleApplicationError(initialError)
 
-			st := status.Convert(handledError)
-			Expect(st.Code()).To(Equal(codes.Internal))
-			Expect(st.Message()).To(Equal(initialError.Error()))
-			Expect(st.Details()).To(HaveLen(1))
-		})
-	})
+	st := status.Convert(handledError)
+	assert.Equal(t, codes.Internal, st.Code())
+	assert.Equal(t, initialError.Error(), st.Message())
+	assert.Equal(t, 1, len(st.Details()))
+}
 
-	Describe("Creating an Application Error", func() {
-		It("should create an application error and set the message and code field values", func() {
-			message := "error message"
-			code := codes.AlreadyExists
+func TestCreateAnApplicationError(t *testing.T) {
+	message := "error message"
+	code := codes.AlreadyExists
 
-			applicationError := handling.NewApplicationError(message, code)
+	applicationError := handling.NewApplicationError(message, code)
 
-			Expect(applicationError).NotTo(BeNil())
-			Expect(applicationError.Error()).To(Equal(message))
-			Expect(applicationError.GRPCStatus().Message()).To(Equal(message))
-			Expect(applicationError.GRPCStatus().Code()).To(Equal(code))
-		})
-	})
-})
+	require.NotNil(t, applicationError)
 
-func TestHandling(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Handling Application Errors Suite")
+	assert.Equal(t, message, applicationError.Error())
+	assert.Equal(t, message, applicationError.GRPCStatus().Message())
+	assert.Equal(t, code, applicationError.GRPCStatus().Code())
 }
