+++ internal/server/teacher/find_by_id.go
@@ -1,0 +1,22 @@
+package teacher
+
+import (
+	"context"
+
+	"github.com/upassed/upassed-account-service/internal/handling"
+	"github.com/upassed/upassed-account-service/pkg/client"
+	"google.golang.org/grpc/codes"
+)
+
+func (server *teacherServerAPI) FindByID(ctx context.Context, request *client.TeacherFindByIDRequest) (*client.TeacherFindByIDResponse, error) {
+	if err := request.Validate(); err != nil {
+		return nil, handling.WrapAsApplicationError(err, handling.WithCode(codes.InvalidArgument))
+	}
+
+	teacher, err := server.service.FindByID(ctx, request.GetTeacherId())
+	if err != nil {
+		return nil, err
+	}
+
+	return ConvertToFindByIDResponse(teacher), nil
+}
