+++ taskfile.yml
@@ -13,4 +13,4 @@ tasks:
     test:
         desc: "Run all rests"
         cmds:
-            - ginkgo --cover --succinct ./...
+            - go test -cover ./...
